centrality and clustering plots {qgraph}	R Documentation

Centrality and Clustering plots and tables

Description

These functions can be used to facilitate interpreting centrality and clustering coefficients. The plot functions use ggplot2 (Wickham, 2009). The table functions create a long format table which can easilly be plotted in ggplot2.

Usage

centralityPlot(..., labels, scale = c("z-scores", "raw", "raw0","relative"),
                 include =c("Degree","Strength","OutDegree","InDegree","OutStrength",
                 "InStrength"), theme_bw = TRUE, print = TRUE, verbose = TRUE, 
                 standardized, relative, weighted = TRUE,signed = TRUE, 
                 orderBy = "default", decreasing = FALSE)
clusteringPlot(..., scale = c("z-scores", "raw", "raw0","relative"), labels, 
                  include , signed = FALSE, theme_bw = TRUE, print = TRUE,
                  verbose = TRUE, standardized, relative,orderBy = "default", 
                  decreasing = FALSE)
centralityTable(..., labels, standardized = TRUE,  relative = FALSE, weighted =
                 TRUE, signed = TRUE)
clusteringTable(..., labels,  standardized = TRUE, relative = FALSE, 
                signed = FALSE)

Arguments

...
Objects usuable in the getWmat generic, such as qgraph objects and weights matrices. Can also be lists containing these objects. Graphs in a list will be plotted in the same panel as different lines and graphs in seperate arguments will be plotted in seperate panels.

scale
Scale of the x-axis. "z-scores" to plot standardized coefficients, "raw" to plot raw coefficients, "raw0" to plot raw coefficients while including 0 on the x-axis and "relative" to show values on a relative scale from 0 (lowest) to 1 (highest).

labels
A vector overwriting the labels used. Can be missing.

include
A vector of measures to include. if missing all measures available will be included. Not included by default are "Closeness", "Betweenness", "ExpectedInfluence", "OutExpectedInfluence", and "InExpectedInfluence". Can also be "all" or "All" to include all available centrality measures.

theme_bw
Adds the ggplot2 black and white theme to the plot

print
If TRUE, the plot is sent to the print command and returned invisible, if FALSE the plot is returned normally. Needed to include plots in e.g., pdf files.

verbose
Should messages be printed to the console?

standardized
Logical, should all measures be standardized? Argument is deprecated and will be removed.

relative
Logical, should all measures be scaled relative to the largest value? Argument is deprecated and will be removed.

weighted
Logical, set to FALSE to set all edge weights to 1 or -1

signed
Logical, set to FALSE to make all edge weights absolute

orderBy
String indicating which measure to order by. Can be default (alphabetical), or one of the measures plotted (e.g., "Strength")

decreasing
Logical indicating if the nodes should be ordered increasing or decreasing

Details

Note that under default setting the plot functions show the standardized centrality indices. That is, z-scores instead of raw centrality indices. This is done to allow easier comparison of multiple networks.

Author(s)

Sacha Epskamp <mail@sachaepskamp.com> \& Jolanda Kossakowski

References

H. Wickham. ggplot2: elegant graphics for data analysis. Springer New York, 2009.

[Package qgraph version 1.6.2 Index]