VARglm {qgraph}	R Documentation

Computes a vector autoregressive lag-1 model using GLM

Description

This function computes a VAR model using glm.

Usage

VARglm(x, family, vars, adjacency, icfun = BIC, ...)

Arguments

x
A data frame

family
The family to be used. Defaults to gaussian if data is continuous or binomial if data is binary

vars
Vector of variables to predict. If missing all variables are predicted.

adjacency
Adjacency matrix. If missing full network is estimated

icfun
Information criterium function to be included in the output

...
Arguments used in the icfun

Value

A list containing:

graph
The estimated graph

IC
The information criterium

Author(s)

Sacha Epskamp <mail@sachaepskamp.com

[Package qgraph version 1.6.2 Index]