as.igraph.qgraph {qgraph}	R Documentation

Converts qgraph object to igraph object.

Description

This function converts the output of qgraph to an 'igraph' object that can be used in the igraph package (Csardi & Nepusz, 2006)

Usage

## S3 method for class 'qgraph'
as.igraph(object, attributes = TRUE)

Arguments

object
A "qgraph" object

attributes
Logical, should graphical attributes also be transferred?

Author(s)

Sacha Epskamp <mail@sachaepskamp.com>

References

Csardi G, Nepusz T (2006). The igraph software package for complex network research, InterJournal, Complex Systems 1695. http://igraph.sf.net

[Package qgraph version 1.6.1 Index]