averageLayout {qgraph}	R Documentation

Computes an average layout over several graphs

Description

This function can be used to compute a joint layout over several graphs.

Usage

averageLayout(..., layout = "spring", repulsion = 1, layout.par)

Arguments

...
Multiple graph objects such as qgraph objects or weights matrices.

layout
Same as in qgraph

repulsion
The repulsion parameter as used in qgraph.

layout.par
Same as in qgraph

Value

A layout matrix

Author(s)

Sacha Epskamp <mail@sachaepskamp.com>

[Package qgraph version 1.6.4 Index]