mat2vec {qgraph}	R Documentation

Weights matrix to vector

Description

Converts a weights matrix to a vector of weights. If the matrix is symmetrical only upper triangle values are returned in the vector.

Usage

mat2vec(x, diag = FALSE, tol = 1e-10)

Arguments

x
A weights matrix

diag
Logical: should diagonal values be included?

tol
Tolerance level

Author(s)

Sacha Epskamp <mail@sachaepskamp.com>

[Package qgraph version 1.6.2 Index]