mutualInformation {qgraph}	R Documentation

Computes the mutual information between nodes

Description

Computes the mutual information from one node to all other nodes, or between sets of nodes.

Usage

mutualInformation(ggm, from, to = "all", covMat)

Arguments

ggm
Partial correlation network. Can be missing if 'covMat' is supplied.

from
Integer vector corresponding to one set of nodes. Defaults to all nodes.

to
Integer vector corresponding to another set of nodes, or 'all' to compute the mutual information of each node to all other nodes.

covMat
Variance-covariance matrix. Can be missing if 'ggm' is supplied.

Author(s)

Sacha Epskamp

[Package qgraph version 1.6.2 Index]