plot.qgraph {qgraph}	R Documentation

Plot method for "qgraph"

Description

Plots an object created by qgraph.

Usage

## S3 method for class 'qgraph'
plot(x, ...)

Arguments

x
A "qgraph" object

...
Not used

Details

If the result of qgraph is stored, such as Graph <- qgraph(...), the plot can be recreated in two ways. qgraph(Graph, ...)) reruns qgraph with the same arguments used in the origina call except those restated in the dots. For example qgraph(Graph, shape = "square") will recreate the same plot but now use square nodes instead of circular. plot(Graph) will NOT rerun qgraph but simply plot the qgraph object. This means that now specific graph attributes can be changed before plotting.

More specific, qgraph(Graph) will base the new plot on the Arguments element of the qgraph object and plot(qgraph) will base the new plot on the graphAttributes element of qgraph.

Author(s)

Sacha Epskamp (mail@sachaepskamp.com)

[Package qgraph version 1.6.2 Index]