qgraphAnnotate {qgraph}	R Documentation

Annotates a qgraph object with mouseover tooltips in a HTML file

Description

This function creates a HTML file based on the result from a qgraph call. The HTML file is made using sendplot

Usage

qgraphAnnotate(graph, ..., fromqgraph = c("labels", "nodeNames", "tooltips", "groups"), 
                filename = "qgraph", image.size = "600x600", window.size = image.size, 
                legend = FALSE)

Arguments

graph
Result from qgraph

...
Named vectors with elements to send to the tooltip. The length of each vector should equal the amount of nodes.

fromqgraph
Vector specifying which arguments from qgraph should be used in the tooltips. Only supported for "labels", "nodeNames", "tooltips" and "groups"

filename
String indicating the root name of the file

image.size
String indicating the size of the file in pixels.

window.size
String indicating the size of the image window in pixels.

legend
Logical, should a legend be plotted? Defaults to FALSE

Details

This function uses the sendplot package (Gaile, Shepherd, Sucheston, Bruno and Manly, 2013).

Value

A string indicating the location of the created HTML file, which can be used with browseURL.

Author(s)

Sacha Epskamp <mail@sachaepskamp.com>

References

Daniel P Gaile, Lori A. Shepherd, Lara Sucheston, Andrew Bruno and Kenneth F. Manly (2013). sendplot: Tool for sending interactive plots with tool-tip content.. R package version 4.0.0. http://CRAN.R-project.org/package=sendplot

[Package qgraph version 1.6.2 Index]