qgraphMixed {qgraph}	R Documentation

Plots a mixed graph with both directed and undirected edges.

Description

This function can be used to plot a network in which each node is connected by at most 3 edges; one undirected edge and two directed edges.

Usage

qgraphMixed(undirected, directed, parallel = TRUE, parallelAngle = pi/6, 
        diagUndirected = FALSE, diagDirected = TRUE, ltyUndirected = 1, ltyDirected = 1, 
        curve = 1, ...)

Arguments

undirected
The undirected network weights matrix.

directed
The directed network weights matrix.

parallel
Logical indicating if edges should be plotted parallel or curved.

parallelAngle
See qgraph

diagUndirected
Logical indicating if the diagonal of the undirected edges should be included.

diagDirected
Logical indicating if the diagonal of the directed edges should be included.

ltyUndirected
lty of undirected edges

ltyDirected
lty of directed edges

curve
Curvature of directed edges

...
Arguments sent to qgraph

Author(s)

Sacha Epskamp <mail@sachaepskamp.com>

[Package qgraph version 1.6.1 Index]