smallworldIndex {qgraph}	R Documentation

Small-world index of unweighted graph

Description

Computes the small-world index of an unweighted graph. When the graph is weighted, weights are removed and every nonzero edge weight is set to 1.

Usage

smallworldIndex(x)

Arguments

x
A qgraph object.

Author(s)

Sacha Epskamp <mail@sachaepskamp.com>

References

Watts, D. J., & Strogatz, S. H. (1998). Collective dynamics of 'small-world' networks. nature, 393(6684), 440-442.

[Package qgraph version 1.6.2 Index]